+++ wezterm-mux-server-impl/src/dispatch.rs
@@ -188,6 +188,18 @@ where
                     .await?;
                 stream.flush().await.context("flushing PDU to client")?;
             }
+            Ok(Item::Notif(MuxNotification::WorkspaceRenamed {
+                old_workspace,
+                new_workspace,
+            })) => {
+                Pdu::RenameWorkspace(codec::RenameWorkspace {
+                    old_workspace,
+                    new_workspace,
+                })
+                .encode_async(&mut stream, 0)
+                .await?;
+                stream.flush().await.context("flushing PDU to client")?;
+            }
             Ok(Item::Notif(MuxNotification::ActiveWorkspaceChanged(_))) => {}
             Ok(Item::Notif(MuxNotification::Empty)) => {}
             Err(err) => {